clean up describe, move result objects to appropriate places
