+++ src/iminuit/util.py
@@ -1,6 +1,7 @@
 """iminuit utility functions and classes."""
 import re
 from collections import OrderedDict, namedtuple
+from collections.abc import Iterable
 from . import _repr_html
 from . import _repr_text
 
@@ -19,6 +20,160 @@ class HesseFailedWarning(IMinuitWarning):
     """HESSE failed warning."""
 
 
+class BasicView:
+    """Array-like view of parameter state.
+
+    Derived classes need to implement methods _set and _get to access
+    specific properties of the parameter state."""
+
+    _minuit = None
+    _ndim = 0
+
+    def __init__(self, minuit, ndim=0):
+        self._minuit = minuit
+        self._ndim = ndim
+
+    def __iter__(self):
+        for i in range(len(self)):
+            yield self._get(i)
+
+    def __len__(self):
+        return self._minuit.narg
+
+    def __getitem__(self, key):
+        if isinstance(key, slice):
+            ind = range(*key.indices(len(self)))
+            return [self._get(i) for i in ind]
+        i = key if isinstance(key, int) else self._minuit._var2pos[key]
+        if i < 0:
+            i += len(self)
+        if i < 0 or i >= len(self):
+            raise IndexError
+        return self._get(i)
+
+    def __setitem__(self, key, value):
+        self._minuit._copy_state_if_needed()
+        if isinstance(key, slice):
+            ind = range(*key.indices(len(self)))
+            if _ndim(value) == self._ndim:  # basic broadcasting
+                for i in ind:
+                    self._set(i, value)
+            else:
+                if len(value) != len(ind):
+                    raise ValueError("length of argument does not match slice")
+                for i, v in zip(ind, value):
+                    self._set(i, v)
+        else:
+            i = key if isinstance(key, int) else self._minuit._var2pos[key]
+            if i < 0:
+                i += len(self)
+            if i < 0 or i >= len(self):
+                raise IndexError
+            self._set(i, value)
+
+    def __eq__(self, other):
+        return len(self) == len(other) and all(x == y for x, y in zip(self, other))
+
+    def __repr__(self):
+        s = "<{} of Minuit at {:x}>".format(self.__class__.__name__, id(self._minuit))
+        for (k, v) in zip(self._minuit._pos2var, self):
+            s += f"\n  {k}: {v}"
+        return s
+
+
+def _ndim(obj):
+    nd = 0
+    while isinstance(obj, Iterable):
+        nd += 1
+        for x in obj:
+            if x is not None:
+                obj = x
+                break
+        else:
+            break
+    return nd
+
+
+class ValueView(BasicView):
+    """Array-like view of parameter values."""
+
+    def _get(self, i):
+        return self._minuit._last_state[i].value
+
+    def _set(self, i, value):
+        self._minuit._last_state.set_value(i, value)
+
+
+class ErrorView(BasicView):
+    """Array-like view of parameter errors."""
+
+    def _get(self, i):
+        return self._minuit._last_state[i].error
+
+    def _set(self, i, value):
+        self._minuit._last_state.set_error(i, value)
+
+
+class FixedView(BasicView):
+    """Array-like view of whether parameters are fixed."""
+
+    def _get(self, i):
+        return self._minuit._last_state[i].is_fixed
+
+    def _set(self, i, fix):
+        if fix:
+            self._minuit._last_state.fix(i)
+        else:
+            self._minuit._last_state.release(i)
+
+
+class LimitView(BasicView):
+    """Array-like view of parameter limits."""
+
+    def _get(self, i):
+        p = self._minuit._last_state[i]
+        return (
+            p.lower_limit if p.has_lower_limit else -inf,
+            p.upper_limit if p.has_upper_limit else inf,
+        )
+
+    def _set(self, i, args):
+        state = self._minuit._last_state
+        val = state[i].value
+        err = state[i].error
+        # changing limits is a cheap operation, start from clean state
+        state.remove_limits(i)
+        low, high = _normalize_limit(args)
+        if low != -inf and high != inf:  # both must be set
+            state.set_limits(i, low, high)
+            if low == high:
+                state.fix(i)
+        elif low != -inf:  # lower limit must be set
+            state.set_lower_limit(i, low)
+        else:  # lower limit must be set
+            state.set_upper_limit(i, high)
+        # bug in Minuit2: must set parameter value and error again after changing limits
+        if val < low:
+            val = low
+        elif val > high:
+            val = high
+        state.set_value(i, val)
+        state.set_error(i, err)
+
+
+def _normalize_limit(lim):
+    if lim is None:
+        return (-inf, inf)
+    lim = list(lim)
+    if lim[0] is None:
+        lim[0] = -inf
+    if lim[1] is None:
+        lim[1] = inf
+    if lim[0] > lim[1]:
+        raise ValueError("limit " + str(lim) + " is invalid")
+    return tuple(lim)
+
+
 class Matrix(tuple):
     """Matrix data object (tuple of tuples)."""
 
@@ -31,14 +186,14 @@ class Matrix(tuple):
 
     def __str__(self):
         """Return string suitable for terminal."""
-        return repr_text.matrix(self)
+        return _repr_text.matrix(self)
 
     def to_table(self):
         args = []
         for mi in self:
             for mj in mi:
                 args.append(mj)
-        nums = repr_text.matrix_format(*args)
+        nums = _repr_text.matrix_format(*args)
         tab = []
         n = len(self)
         for i, name in enumerate(self.names):
@@ -150,7 +305,7 @@ class FMin:
 
     def __str__(self):
         """Return string suitable for terminal."""
-        return repr_text.fmin(self)
+        return _repr_text.fmin(self)
 
     def _repr_html_(self):
         return _repr_html.fmin(self)
@@ -240,11 +395,11 @@ class Params(list):
             row = [i, name]
             if mes and name in mes:
                 me = mes[name]
-                val, err, mel, meu = repr_text.pdg_format(
+                val, err, mel, meu = _repr_text.pdg_format(
                     mp.value, mp.error, me.lower, me.upper
                 )
             else:
-                val, err = repr_text.pdg_format(mp.value, mp.error)
+                val, err = _repr_text.pdg_format(mp.value, mp.error)
                 mel = ""
                 meu = ""
             row += [
@@ -261,7 +416,7 @@ class Params(list):
 
     def __str__(self):
         """Return string suitable for terminal."""
-        return repr_text.params(self)
+        return _repr_text.params(self)
 
     def _repr_pretty_(self, p, cycle):
         if cycle:
@@ -313,7 +468,7 @@ class MError:
 
     def __str__(self):
         """Return string suitable for terminal."""
-        return repr_text.merrors([self])
+        return _repr_text.merrors([self])
 
     def _repr_pretty_(self, p, cycle):
         if cycle:
@@ -332,7 +487,7 @@ class MErrors(OrderedDict):
 
     def __str__(self):
         """Return string suitable for terminal."""
-        return repr_text.merrors(self.values())
+        return _repr_text.merrors(self.values())
 
     def _repr_pretty_(self, p, cycle):
         if cycle:
@@ -383,44 +538,22 @@ def make_func_code(params):
     return namedtuple("FuncCode", "co_varnames co_argcount")(params, len(params))
 
 
-def describe(f):
+def describe(callable):
     """Try to extract the function argument names."""
     return (
-        _arguments_from_funccode(f)
-        or _arguments_from_call_funccode(f)
-        or _arguments_from_docstring(f.__call__.__doc__)
-        or _arguments_from_docstring(f.__doc__)
-        or _arguments_from_inspect(f)
-        or None
+        _arguments_from_func_code(callable)
+        or _arguments_from_inspect(callable)
+        or _arguments_from_docstring(callable.__call__.__doc__)
+        or _arguments_from_docstring(callable.__doc__)
     )
 
 
-def _fc_or_c(f):
-    if hasattr(f, "func_code"):
-        return f.func_code
-    if hasattr(f, "__code__"):
-        return f.__code__
-    # bound method and fake function will be None
-    return None
-
-
-def _arguments_from_funccode(f):
-    # Check f.funccode for arguments
-    fc = _fc_or_c(f)
-    if fc is None:
-        return None
-    nargs = fc.co_argcount
-    args = fc.co_varnames[:nargs]
-    if getattr(f, "__self__", None) is not None:
-        args = args[1:]
-    return list(args)
-
-
-def _arguments_from_call_funccode(f):
-    # Check f.__call__.func_code for arguments
-    fc = _fc_or_c(f.__call__)
-    nargs = fc.co_argcount
-    return list(fc.co_varnames[1:nargs])
+def _arguments_from_func_code(obj):
+    # Check (faked) f.func_code; for backward-compatibility with iminuit-1.x
+    if hasattr(obj, "func_code"):
+        fc = obj.func_code
+        nargs = fc.co_argcount
+        return fc.co_varnames[:nargs]
 
 
 def _arguments_from_inspect(f):
@@ -428,13 +561,15 @@ def _arguments_from_inspect(f):
     import inspect
 
     signature = inspect.signature(f)
+    args = []
     for name, par in signature.parameters.items():
         # Variable number of arguments is not supported
         if par.kind is inspect.Parameter.VAR_POSITIONAL:
             return None
         if par.kind is inspect.Parameter.VAR_KEYWORD:
-            return None
-    return list(signature.parameters)
+            break
+        args.append(name)
+    return args
 
 
 def _arguments_from_docstring(doc):
@@ -478,3 +613,7 @@ def _arguments_from_docstring(doc):
         ret = ret[1:]
 
     return ret
+
+
+def _guess_initial_step(val):
+    return 1e-2 * val if val != 0 else 1e-1  # heuristic
